+++ pyinterpolate/kriging/models/point/simple_kriging.py
@@ -5,14 +5,12 @@ Authors
 -------
 1. Szymon Moliński | @SimonMolinsky
 
-TODO
-----
-* log k, predicted, dataset
 """
 # Python core
 from typing import List, Union, Tuple
 
 # Core calculation and data visualization
+import logging
 import numpy as np
 
 # Pyinterpolate
@@ -76,13 +74,14 @@ def simple_kriging(
     RunetimeError
         Singularity matrix in a Kriging system.
     """
-
+    logging.info("SIMPLE KRIGING | Operation starts")
     k, predicted, dataset = get_predictions(theoretical_model,
                                             known_locations,
                                             unknown_location,
                                             neighbors_range,
                                             no_neighbors,
                                             use_all_neighbors_in_range)
+    logging.info("SIMPLE KRIGING | Weights are prepared")
 
     try:
         output_weights = solve_weights(predicted, k, allow_approximate_solutions)
@@ -98,6 +97,8 @@ def simple_kriging(
     sigma = np.matmul(output_weights.T, k)
 
     if sigma < 0:
+        logging.info("SIMPLE KRIGING | Variance error lower than zero, NaN is returned")
         return [zhat, np.nan, unknown_location[0], unknown_location[1]]
 
+    logging.info("SIMPLE KRIGING | Process ends")
     return [zhat, sigma, unknown_location[0], unknown_location[1]]