+++ changelog.rst
@@ -12,6 +12,8 @@ Changes by date
 **version 0.3.7**
 
 * (enhancement) added logging to Poisson Kriging ATP process,
+* (enhancement) added logging to Simple Kriging process,
+*
 
 2023-01-16
 ----------